requirements.txt
imaplib2  
email  
yagmail
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
tavily-search  
SQLAlchemy
tenacity
